requirements.txt
asgiref==3.11.1
crispy-tailwind==1.0.3
Django==5.2.12
django-crispy-forms==2.6
django-stubs-ext==5.2.9
django-widget-tweaks==1.5.1
djangorestframework==3.16.1
librt==0.8.1
mypy_extensions==1.1.0
pathspec==1.0.4
pillow==12.1.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
tomli==2.4.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
django-environ == 0.13.0
